Remove deprecated decorateWidget() call
Docs do not immediately indicate what should replace this:
- https://meta.discourse.org/t/preparing-your-community-for-behind-the-scenes-header-changes/316549
- https://meta.discourse.org/t/upcoming-header-changes-preparing-themes-and-plugins/296544?page=2

Empirically, this call silently fails with an error, and the theme seems
to do what we want even without it. In 3.4.0-beta1, the new header was
enabled by default, and it worked fine. In beta2, the old header is
gone, so is the warning banner, and it still works fine.

As far as I can see, all that the obsolete call does is generate an
error and then move on to the createWidget() call which actually does
work.

Note that out of the box we have the New Topic button in *most* of the
right places, in fact this theme hides it in some cases. The theme adds
it on the topic (post) page, at the top *and* at the bottom. The
official "New Topic button" theme *does not* add it at the bottom.